+++ src/php/Triniti/Schemas/Users/Mixin/CreateUser/CreateUser.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Triniti\Schemas\Users\Mixin\CreateUser;
-
-use Gdbots\Schemas\Ncr\Mixin\CreateNode\CreateNodeV1;
-
-interface CreateUser extends CreateNodeV1
-{
-}