+++ src/php/Triniti/Schemas/Users/Mixin/GetUserBatchRequest/GetUserBatchRequest.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Triniti\Schemas\Users\Mixin\GetUserBatchRequest;
-
-use Gdbots\Schemas\Ncr\Mixin\GetNodeBatchRequest\GetNodeBatchRequestV1;
-
-interface GetUserBatchRequest extends GetNodeBatchRequestV1
-{
-}