+++ src/php/Triniti/Schemas/Users/Mixin/GetUserBatchRequest/GetUserBatchRequestV1.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Triniti\Schemas\Users\Mixin\GetUserBatchRequest;
-
-interface GetUserBatchRequestV1 extends GetUserBatchRequest
-{
-}