+++ src/php/Triniti/Schemas/Users/Mixin/GetUserBatchResponse/GetUserBatchResponse.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Triniti\Schemas\Users\Mixin\GetUserBatchResponse;
-
-use Gdbots\Schemas\Ncr\Mixin\GetNodeBatchResponse\GetNodeBatchResponseV1;
-
-interface GetUserBatchResponse extends GetNodeBatchResponseV1
-{
-}