+++ src/php/Triniti/Schemas/Users/Mixin/GetUserHistoryResponse/GetUserHistoryResponse.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Triniti\Schemas\Users\Mixin\GetUserHistoryResponse;
-
-use Gdbots\Schemas\Pbjx\Mixin\GetEventsResponse\GetEventsResponseV1;
-
-interface GetUserHistoryResponse extends GetEventsResponseV1
-{
-}