+++ src/php/Triniti/Schemas/Users/Mixin/GetUserRequest/GetUserRequest.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Triniti\Schemas\Users\Mixin\GetUserRequest;
-
-use Gdbots\Schemas\Ncr\Mixin\GetNodeRequest\GetNodeRequestV1;
-
-interface GetUserRequest extends GetNodeRequestV1
-{
-}