+++ src/php/Triniti/Schemas/Users/Mixin/GetUserResponse/GetUserResponseV1.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Triniti\Schemas\Users\Mixin\GetUserResponse;
-
-interface GetUserResponseV1 extends GetUserResponse
-{
-}