+++ src/php/Triniti/Schemas/Users/Mixin/GrantRolesToUser/GrantRolesToUser.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Triniti\Schemas\Users\Mixin\GrantRolesToUser;
-
-use Gdbots\Schemas\Pbjx\Mixin\Command\CommandV1;
-
-interface GrantRolesToUser extends CommandV1
-{
-}