+++ src/php/Triniti/Schemas/Users/Mixin/RevokeRolesFromUser/RevokeRolesFromUser.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Triniti\Schemas\Users\Mixin\RevokeRolesFromUser;
-
-use Gdbots\Schemas\Pbjx\Mixin\Command\CommandV1;
-
-interface RevokeRolesFromUser extends CommandV1
-{
-}