+++ src/php/Triniti/Schemas/Users/Mixin/SearchUsersRequest/SearchUsersRequestV1.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Triniti\Schemas\Users\Mixin\SearchUsersRequest;
-
-interface SearchUsersRequestV1 extends SearchUsersRequest
-{
-}