+++ src/php/Triniti/Schemas/Users/Mixin/SearchUsersResponse/SearchUsersResponse.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Triniti\Schemas\Users\Mixin\SearchUsersResponse;
-
-use Gdbots\Schemas\Ncr\Mixin\SearchNodesResponse\SearchNodesResponseV1;
-
-interface SearchUsersResponse extends SearchNodesResponseV1
-{
-}