+++ src/php/Triniti/Schemas/Users/Mixin/UpdateUser/UpdateUser.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Triniti\Schemas\Users\Mixin\UpdateUser;
-
-use Gdbots\Schemas\Ncr\Mixin\UpdateNode\UpdateNodeV1;
-
-interface UpdateUser extends UpdateNodeV1
-{
-}