+++ src/php/Triniti/Schemas/Users/Mixin/User/User.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Triniti\Schemas\Users\Mixin\User;
-
-use Gdbots\Schemas\Ncr\Mixin\Node\NodeV1;
-
-interface User extends NodeV1
-{
-}