+++ src/php/Triniti/Schemas/Users/Mixin/UserDeleted/UserDeleted.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Triniti\Schemas\Users\Mixin\UserDeleted;
-
-use Gdbots\Schemas\Ncr\Mixin\NodeDeleted\NodeDeletedV1;
-
-interface UserDeleted extends NodeDeletedV1
-{
-}