+++ src/php/Triniti/Schemas/Users/Mixin/UserDeleted/UserDeletedV1.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Triniti\Schemas\Users\Mixin\UserDeleted;
-
-interface UserDeletedV1 extends UserDeleted
-{
-}