+++ src/php/Triniti/Schemas/Users/Mixin/UserRolesGranted/UserRolesGranted.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Triniti\Schemas\Users\Mixin\UserRolesGranted;
-
-use Gdbots\Schemas\Pbjx\Mixin\Event\EventV1;
-
-interface UserRolesGranted extends EventV1
-{
-}