+++ src/php/Triniti/Schemas/Users/Mixin/UserRolesRevoked/UserRolesRevokedV1.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Triniti\Schemas\Users\Mixin\UserRolesRevoked;
-
-interface UserRolesRevokedV1 extends UserRolesRevoked
-{
-}