+++ src/php/Triniti/Schemas/Users/Mixin/UserUpdated/UserUpdated.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Triniti\Schemas\Users\Mixin\UserUpdated;
-
-use Gdbots\Schemas\Ncr\Mixin\NodeUpdated\NodeUpdatedV1;
-
-interface UserUpdated extends NodeUpdatedV1
-{
-}