+++ CHANGELOG.md
@@ -3,3 +3,24 @@
 
 ## v1.0.0
 * Initial version.
+* __Schemas:__
+  * `triniti:pages:mixin:create-page`
+  * `triniti:pages:mixin:delete-page`
+  * `triniti:pages:mixin:get-page-batch-request`
+  * `triniti:pages:mixin:get-page-batch-response`
+  * `triniti:pages:mixin:get-page-history-request`
+  * `triniti:pages:mixin:get-page-history-response`
+  * `triniti:pages:mixin:get-page-request`
+  * `triniti:pages:mixin:get-page-response`
+  * `triniti:pages:mixin:page`
+  * `triniti:pages:mixin:page-created`
+  * `triniti:pages:mixin:page-deleted`
+  * `triniti:pages:mixin:page-published`
+  * `triniti:pages:mixin:page-unpublished`
+  * `triniti:pages:mixin:page-updated`
+  * `triniti:pages:mixin:publish-page`
+  * `triniti:pages:mixin:search-pages-request`
+  * `triniti:pages:mixin:search-pages-response`
+  * `triniti:pages:mixin:unpublish-page`
+  * `triniti:pages:mixin:update-page`
+